requirements.txt
streamlit
langchain
pinecone-client
sentence-transformers
PyPDF2
huggingface_hub
